Разбор PosData

+++ EGTS/ServiceLayerData/ServiceDataRecords/Subrecords/SubrecordData/TeledataServiceRecords/PosDataSubrecord.cs
@@ -21,12 +21,12 @@ namespace EGTS.ServiceLayer.TeledataService
 
         public float Latitude { get; set; }
         public float Longitude { get; set; }
-        public ushort Speed { get; set; }
+        public float Speed { get; set; }
         public ushort Direction { get; set; }
-        public uint Odometer { get; set; }
+        public float Odometer { get; set; }
         public byte DigitalInputs { get; set; }
         public byte Source { get; set; }
-        public uint Altitude { get; set; }
+        public int Altitude { get; set; }
         public bool Valid { get; set; }
         public bool Actual { get; set; }
         public bool Moving { get; set; }
